pyproject.toml
[tool.poetry]
name = "lineal-rugby"
version = "0.1.0"
description = ""
authors = ["Andy Sprague <andy.sprague44@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
aws-cdk-lib = "^2.151.0"
boto3 = "^1.34.157"
cdk-nag = "^2.28.175"
aws-lambda-powertools = "^2.43.0"

[tool.poetry.group.dev.dependencies]
black = "*"
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
